Former-commit-id: 5a3fea98eb2bb9c13d6b32e592d4bbfec102eff2
Former-commit-id: 619416cf0951727f6d2e031e4f3346c87e961e92
Former-commit-id: 2791d036367bbc0434c6d9533b333037df4e1325

+++ vizia_api_bindings/python/ViziaDoomGamePython.h
@@ -23,7 +23,7 @@ namespace Vizia {
     public:
         struct PythonState
         {
-            int number;
+            unsigned int number;
             object imageBuffer;
             object gameVariables;
             PythonState(int n, object buf, object v ):number(n),imageBuffer(buf),gameVariables(v){}
@@ -36,8 +36,8 @@ namespace Vizia {
         PythonState getState();
         boost::python::list getLastAction();
         object getGameScreen();
-        float makeAction(boost::python::list &action);
-        float makeAction(boost::python::list &action, unsigned int tics);
+        double makeAction(boost::python::list &action);
+        double makeAction(boost::python::list &action, unsigned int tics);
 
     private:
